Enabled scripting and added Docker image files
First, a large number of changes were made to the way some of the paths
and top-level files are handled in Python to ensure that all of the
scripts can be run as scripts from a terminal. Second, Dockerfile was
added to generate an image that can run the entire repo for a completely
contained manuscript reproduction.