Initial stab at question
Initial effort in modeling difference between Experiment 1a and
Experiment 1b. Intriguingly, it appears that, at least at higher levels,
it is the case that the AI model predicts better thresholds in Exp 1b
than in Exp 1a. This is a really fascinating result. Need to look
further into this and make sure it is not artifactual.